First pass at implementing suggested changes
